The Base Layout Update
Want to have the same base layout across pages on your website? This is now much easier to achieve than ever!

Simply set the metadata key "layout" to the path to your layout file, relative to where you run StaticPHP.

StaticPHP will load in this file, combine the content of the current page into the base layout (based on a customisible placeholder you can set in the metadata of the base layout), and process it as one bigger page.

Currently, this functionality is only available for PHP files, but I plan to extend this feature when I add support for other content types, such as Markdown and standard HTML.